storage: expose persist's inspect_shard at the controller level
We'll be thoughtful about making unnecessary changes, but the **output
of this method needs to be gated from users**, so that it's not subject
to our backward compatibility guarantees.

TODO: Ideally this would return `impl Serialize` so the caller can do
with it what they like, but that doesn't work in traits yet. The
workaround (an associated type) doesn't work because persist doesn't
want to make the type public. In the meantime, move the `serde_json`
call from the single user into this method.