Implement auto_index_selects for sqllogictest
This commit provides an implementation of the auto_index_selects option of sqllogictest. The implementation consists of obtaining outcomes for running a SELECT query in the SLT file unchanged and for running the same query created as an indexed view. The outcomes are contrasted, and it is a test failure if the outcomes obtained are not of the same type. One important caveat is that queries with column ambiguity in SLT files may lead to view definitions that error. This problem will be handled by a subsequent commit.