sql: implement minimal support for pg_cancel_backend
pg_cancel_backend is a side-effecting function that cancels the
identified connection. Because Materialize's compute layer cannot handle
side-effecting functions, we introduce a new code path in the planner
that sniffs out the common form of calls to these functions (`SELECT
side_effecting_function()`) and handles them entirely in the adapter
layer.