adapter: unify dataflow IDs
This commit unifies the dataflow IDs exposed by the introspection views
by renaming the previous `mz_dataflows.local_id` field to `id` and
dropping the previous `id` field. All dataflow foreign keys now
reference `mz_dataflows.id`.

The purpose of this commit is to remove the confusion about which of the
two exposed IDs is the correct one for joining relations, and what is
the difference between the two IDs.