persist: fix bug with stats on shards with all no-stats columns
We skip stats on certain ScalarTypes, including MzAclItem. If every
column in a struct is skipped, we end up with an empty struct. This
resulted in a len-mismatch error at validate time because we were
defaulting empty structs to len 0. Instead, manually maintain the length
of columnar structs during construction.