persist: fix bug in StateDiff::validate_roundtrip
It was using the "to" applier_version, hostname, and walltime_ms, which
meant the diff didn't cleanly apply. We weren't noticing it before 19954
because of a bug that meant we weren't applying diffs for these fields.
I'm honestly a little confused why we haven't been noticing it since
then. This got caught because of a CI failure in #20013 with
`test-compute-reconciliation-no-errors`, but I'm honestly baffled how
that test is passing right now on main.