adapter: Refactor privilege checks for sys privs (#20187)
This commit refactors the RBAC privilege check code to use the type
`SystemObjectId` instead of `ObjectId`. This has no effect on the actual
logic, but is needed in preparation for checking system privileges. This
is a separate commit because it's such a large amount of lines changed
that it would distract from other changes if it was included in another
commit.

Works towards resolving #19997