storage: refactor resumption frontier handling
This PR clears up some things around the resumption frontier handling in
ingestions. First, it separates the ingestion's `as_of` frontier, which
is the frontier beyond which all inputs must accumulate to correct
collections and the resumption frontier which is the frontier beyond
which the updates *must* be produced.

What this means is that if as_of < resumption_frontier then a snapshot
is not necessary to be produced and it is otherwise.

This PR also deletes the resumption frontier calculator machinery and
the resumption operator. In their place we simply observe the combined
upper frontiers of all the rendered persist sinks.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>