Adjust command history leak test to avoid flakiness
This commit adjusts the bounds for acceptable dataflow counts in the compute command histories of the controller and replica in the cluster test for issue #15531, namely ensuring that we resolved a peek leak in the command history. The bounds are relaxed from 1 to account for a maximum of 2 live dataflows after history reduction plus a fudge factor to account for when reduction actually takes place.