adapter: correctly compute dependencies for materialized views
This commit fixes a bug introduced by #19648, in which we stopped
tracking the dependencies of a materialized view correctly.

The root cause was that there are two notions of dependencies tracked by
the system:

  1. An object required while planning an `mz_sql_parser::ast::Statement`.
  2. An object required to determine the timestamp of a `MirRelationExpr` in
     a plan.

Pre PR #19648, the code used (1) in a place where (2) was expected.
Post PR #19648, the code used (2) in a place where (1) was expected.
This commit uses (1) where (1) is expected and (2) where (2) is
expected.

A future commit will make this easier to understand.

Fix #20315.