storage/postgres: use keepalive messages to drive progress
In the absence of transactions the only available piece of information
in the replication stream are keepalive messages. Keepalive messages are
documented[1] to contain the current end of WAL on the server. That is a
useless number when it comes to progress tracking because there might be
pending messages at LSNs between the last received commit_lsn and the
current end of WAL.

Fortunatelly for us, the documentation for PrimaryKeepalive messages is
wrong and it actually contains the last *sent* LSN[2]. Here sent doesn't
necessarily mean sent over the wire, but sent to the upstream process
that is handling producing the logical stream. Therefore, if we receive
a keepalive with a particular LSN we can be certain that there are no
other replication messages at previous LSNs, because they would have
been already generated and received. We therefore connect the keepalive
messages directly to our capability.

[1]: https://www.postgresql.org/docs/15/protocol-replication.html#PROTOCOL-REPLICATION-START-REPLICATION
[2]: https://www.postgresql.org/message-id/CAFPTHDZS9O9WG02EfayBd6oONzK%2BqfUxS6AbVLJ7W%2BKECza2gg%40mail.gmail.com

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>